Store wellbeing as percent and normalize input
Migrate wellbeing values to a percent (0-100) representation and normalize inputs.

- Added migration to alter pins.wellbeing to DECIMAL(5,2), convert existing 0-10 values to percent (ROUND((wellbeing/10)*100,2)), update slider question configs, and enable use_for_color for the wellbeing question.
- Updated helpers.php to treat wellbeing as a float.
- Updated pins.php to normalize incoming wellbeing answers via a new normalize_percent() helper that clamps numeric input to 0-100 and rounds to two decimals.

This ensures stored wellbeing values use a consistent percent format and incoming data is validated/coerced accordingly.

+++ helpers.php
@@ -9,7 +9,7 @@ function normalize_pin_row(array $row): array
     $row['position_x'] = floatval($row['position_x']);
     $row['position_y'] = floatval($row['position_y']);
     $row['position_z'] = floatval($row['position_z']);
-    $row['wellbeing'] = intval($row['wellbeing']);
+    $row['wellbeing'] = floatval($row['wellbeing']);
     $row['approved'] = intval($row['approved']);
     $row['group_key'] = $row['group_key'] ?? null;
     return $row;
